refactor(Lotto): 패키지 변경
