+++ src/main/java/lotto/domain/LottoMaker.java
@@ -8,7 +8,6 @@ import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import lotto.Lotto;
 
 public class LottoMaker {
