+++ src/main/java/lotto/service/LottoService.java
@@ -1,7 +1,7 @@
 package lotto.service;
 
 import java.util.List;
-import lotto.Lotto;
+import lotto.domain.Lotto;
 import lotto.domain.BonusNumber;
 import lotto.domain.LottoMarket;
 import lotto.domain.NumbersComparator;
