+++ src/main/java/lotto/view/OutputView.java
@@ -2,7 +2,7 @@ package lotto.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import lotto.Lotto;
+import lotto.domain.Lotto;
 
 public class OutputView {
